requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
pymongo==4.8.0
aiohttp==3.9.1
dnspython==2.6.1